requirements.txt
alembic==1.17.2
importlib_metadata==8.7.0
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
sqlparse==0.5.4
tabulate==0.9.0
typing_extensions==4.15.0
zipp==3.23.0
